nlquotes: pick a random-er quote
We want to pick a _random_ quote from the database, but searches only
give us the first 10 entries by default. Therefore, we take the total
count of entries and pick a random entry. Then, if that entry is on a
page past the first one, we fetch that page and then pick a random quote
from that entry.